requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.29.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.5
markovify==0.9.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.2
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.0
Wikipedia-API==0.8.1
